pyproject.toml
[project]
name = "aws-s3-datalake"
version = "0.1.0"
description = ""
authors = [
    {name = "Ricardo Marques",email = "dataricardom@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "boto3 (>=1.42.65,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pandas (>=3.0.1,<4.0.0)",
    "pyarrow (>=23.0.1,<24.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
